feat: add installment service with due date validation
Added a new file implementing the installment service. The Service struct includes the InstallmentRepository interface. A Create method was implemented to add new installments, which includes a DueDay validation to prevent the creation of installments with expired due dates. In addition, the services container was also updated to include the new InstallmentService.